pyproject.toml
[tool.poetry]
name = "bls-bible"
version = "0.1.0"
description = "The BLS Bible: The keeper of truth"
authors = ["Thomas Presto", "Cody Martin"]
license = "GNU GENERAL PUBLIC LICENSE"
readme = "README.md"
packages = [
	{include = "bls_bible"}
]
homepage = "https://github.com/blacklanternsecurity/bls-bible"
repository = "https://github.com/blacklanternsecurity/bls-bible"

[tool.poetry.scripts]
bls-bible="bls_bible.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "4.9.3"
Brotli = "1.0.9"
brotlicffi = "1.0.9.2"
certifi = "2022.9.14"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
click = "7.1.2"
cssselect2 = "0.6.0"
Flask = "1.1.4"
fonttools = {version = "4.37.2", extras = ["woff"]}
html5lib = "1.1"
idna = "3.4"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
markdown2 = "2.4.3"
MarkupSafe = "2.0.1"
objectpath = "0.6.1"
pbr = "5.10.0"
Pillow = "9.2.0"
pycparser = "2.21"
pydyf = "0.2.0"
pyphen = "0.13.0"
python-gitlab = "2.6.0"
PyYAML = "5.3.1"
requests-toolbelt = "0.9.1"
requests = "2.28.1"
six = "1.16.0"
soupsieve = "2.3.2.post1"
testresources = "2.0.1"
tinycss2 = "1.1.1"
urllib3 = "1.26.12"
weasyprint = "54.3"
webencodings = "0.5.1"
Werkzeug = "1.0.1"
zopfli = "0.2.1"
prettytable = "^3.4.1"
ansible-runner = "^2.2.1"
omegaconf = "^2.2.3"
termcolor = "^2.0.1"
GitPython = "^3.1.27"
schedule = "^1.1.0"
argcomplete = "^2.0.0"
Flask-RESTful = "^0.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
